Refactoring change id type to long

+++ src/main/java/com/example/demo/repository/JobRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobRepo extends ListCrudRepository<Job,Integer> {
+public interface JobRepo extends ListCrudRepository<Job,Long> {
 }
